issue 27
modified the size of the raid icons on the player frames. This is a work in progress, as I need to make sure they all appear as they should, including secondary frames. I also need to make the same modification on the party frames

I also removed the pet code from the party frames, temporarily. This is so that I can release RC7 prior to finishing the pet code.