issue 12
Also, I took the liberty of fixing the position of the VehicleSeatIndicator, BattlefieldMinimap, and DurabilityFrame. All of which had potential overlap at certain resolutions. This may still be an issue at higher sui scales, but I doubt people will be looking to make sui bigger anyway.

I also added feedback for the /sui party command so that it tells you when it unlocks and locks the frame. Also, I made sure that the lock mechanic doesn't persist through a reload or zone, and that it gets locked when you use /sui party reset.