Update issue 2 Status: Resolved Labels: Milestone-2.6.0-RC8
Added a focus frame to the list of player frames available. This new frame allows the user to move it around, by using the /sui focus [lock|unlock|reset] command. Also added the ability to enable and disable buffs on this frame using /sui auras focus. These features may need to be thoroughly tested and debugged prior to release.